+++ instrumentation/opentelemetry-instrumentation-asyncpg/src/opentelemetry/instrumentation/asyncpg/__init__.py
@@ -34,6 +34,7 @@ API
 ---
 """
 
+import re
 from typing import Collection
 
 import asyncpg
@@ -99,6 +100,7 @@ class AsyncPGInstrumentor(BaseInstrumentor):
         super().__init__()
         self.capture_parameters = capture_parameters
         self._tracer = None
+        self._leading_comment_remover = re.compile(r"^/\*.*?\*/")
 
     def instrumentation_dependencies(self) -> Collection[str]:
         return _instruments
@@ -135,7 +137,8 @@ class AsyncPGInstrumentor(BaseInstrumentor):
         name = args[0] if args[0] else params.get("database", "postgresql")
 
         try:
-            name = name.split()[0]
+            # Strip leading comments so we get the operation name.
+            name = self._leading_comment_remover.sub("", name).split()[0]
         except IndexError:
             name = ""